Added srollRecursively method in the Utility class and updated the test case.
srollRecursively method has a input param which is String xpath. it will get the list of elements based on the xpath and then store the size of it which will be Initial count of elements. Next it will move to the last element and wait until the content in the DOM is loaded. Now again it will be the size of the elements with the given xpath and validate whether the initial and iterated element size is equal or not. if condition is satisfied the loop is ended. else the initial count is updated with the iterated element size and scroll's down to the last element and loop continues.
-------------------------------------------

Updated the test case to demonstrate the usage of keyboardActions and scrollRecursively. The Method keyboardActions is used to mimic the keyboard actions like entering text into text field and hitting enter button. I have declared the same in the testcase.