Added methods loadPropertyFile, getPropertyValue, openNewTab,, goTo
loadPropertyFile(String propertyFilePath) : This method will take the file path of the property file and configure it accordingly.

getPropertyValue(String key): This method is used to retrieve the value from the property file by taking the key as input param.

openNewTab(): By this method we can open new tab in the active browser. It executes a javascript to achieve it

goTo(String url): This method navigates to input url if the browser is already open and running.